caseDicts/postProcessing: Added triSurfaceAverage packaged function object
This is a packaged function object that conveniently computes averages
of fields on tri-surfaces. It can be executed on the command line as
follows:

    foamPostProcess -func "triSurfaceAverage(name=mid.obj, p)"

This will compute the average of the field "p" on a surface file in
"constant/geometry/mid.obj".

The calculation could also be done at run-time by adding the following
entry to the functions section of the system/controlDict

    #includeFunc triSurfaceAverage(name=mid.obj, p)