lagrangian: Added support for mesh distribution
This change makes particles distribute between processors along with
their containing cells. This might be necessary, for example, if the
chemistry solution is being load balanced, or the cells are being
continually re-decomposed due to dynamic refinement.

Whilst this adds load balancing *with* lagrangian, it does not
facilitate load balancing *of* lagrangian. That would also require the
implementation of a lagranian CPU load that the distributor would then
attempt to balance.